--- backend.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.31903.59
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "backend", "backend\backend.csproj", "{7C1D013F-A76B-4370-86CD-6EE47DFEDF89}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SpecFlowBackend", "SpecFlowBackend\SpecFlowBackend\SpecFlowBackend.csproj", "{B66163AF-3EBC-477B-971D-927A3C608353}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7C1D013F-A76B-4370-86CD-6EE47DFEDF89}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7C1D013F-A76B-4370-86CD-6EE47DFEDF89}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7C1D013F-A76B-4370-86CD-6EE47DFEDF89}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7C1D013F-A76B-4370-86CD-6EE47DFEDF89}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B66163AF-3EBC-477B-971D-927A3C608353}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B66163AF-3EBC-477B-971D-927A3C608353}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B66163AF-3EBC-477B-971D-927A3C608353}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B66163AF-3EBC-477B-971D-927A3C608353}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
