+++ backend/LaakAPI.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "backend", "LaakAPI\backend.csproj", "{3899AFBE-3DA7-48FA-A3A1-012282190FB9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SpecFlowBackend", "LaakAPI.Tests\SpecFlowBackend.csproj", "{65D2F486-CE35-44B9-A04D-0C8A8A396322}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{3899AFBE-3DA7-48FA-A3A1-012282190FB9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3899AFBE-3DA7-48FA-A3A1-012282190FB9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3899AFBE-3DA7-48FA-A3A1-012282190FB9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3899AFBE-3DA7-48FA-A3A1-012282190FB9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{65D2F486-CE35-44B9-A04D-0C8A8A396322}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{65D2F486-CE35-44B9-A04D-0C8A8A396322}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{65D2F486-CE35-44B9-A04D-0C8A8A396322}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{65D2F486-CE35-44B9-A04D-0C8A8A396322}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
